+++ src/Dep.Runtime.Logging.Providers.NLog/GlobalSuppressions.cs
@@ -1,0 +1,8 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Microsoft.Design", "BaseMethodViolations:Exception(s) thrown but not declared in base method", Justification = "<挂起>", Scope = "member", Target = "~M:NetowlsStudio.Dep.Runtime.Logging.Providers.Configuration.XmlConfigurationSetup.Get~NLog.Config.LoggingConfiguration")]